change to bigint

+++ Database/init/create.sql
@@ -13,6 +13,6 @@ CREATE TABLE public.passkey_credentials (
   id varchar NOT null,
   user_id uuid NOT null references users(id),
   public_key bytea not null,
-  sign_count integer not null,
+  sign_count bigint not null,
   CONSTRAINT passkey_credentials_pk PRIMARY KEY (id)
 );
